+++ spec/network/error_spec.rb
@@ -9,7 +9,7 @@ module Ferrum
         expect(network.idle?).to be_falsey
 
         # FIXME: Hack to wait for content in the browser
-        Ferrum.with_attempts(errors: RuntimeError, max: 10, wait: 0.1) do
+        Utils::Attempt.with_retry(errors: RuntimeError, max: 10, wait: 0.1) do
           page.at_xpath("//h1[text() = 'Canceled']") || raise("Node not found")
         end