+++ spec/unit/process_spec.rb
@@ -5,7 +5,7 @@ module Ferrum
     describe Process do
       subject { Browser.new(port: 6000, host: "127.0.0.1") }
 
-      unless Ferrum.windows?
+      unless Utils::Platform.windows?
         it "forcibly kills the child if it does not respond to SIGTERM" do
           allow(::Process).to receive_messages(spawn: 5678)
           allow(::Process).to receive(:wait).and_return(nil)